refactor(#5031): extract tmux relay coordination
